+++ data/src/main/java/com/github/dodobest/data/di/UpbitDataModule.kt
@@ -1,4 +1,4 @@
-package com.github.dodobest.data.hilt
+package com.github.dodobest.data.di
 
 import com.github.dodobest.data.Constant
 import com.github.dodobest.data.data.UpbitAPI
